requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.1
cssselect==1.2.0
cycler==0.12.1
datasets==2.21.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
Django==5.1.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.114.0
fastapi-cli==0.0.5
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
itemadapter==0.9.0
itemloaders==1.3.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.26.1
packaging==24.1
pandas==2.2.2
parsel==1.9.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.1
prompt_toolkit==3.0.47
Protego==0.3.1
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.7.0
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
rich==13.8.1
safetensors==0.4.5
Scrapy==2.11.2
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.38.5
sympy==1.13.2
timm==1.0.9
tldextract==5.1.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==3.0.0
Twisted==24.7.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
w3lib==2.2.1
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.10.0
zope.interface==7.0.3
